+++ crd-generator/api/src/test/java/io/fabric8/crd/example/annotated/AnnotatedSpec.java
@@ -24,7 +24,9 @@ import io.fabric8.generator.annotation.Min;
 import io.fabric8.generator.annotation.Nullable;
 import io.fabric8.generator.annotation.Pattern;
 import io.fabric8.generator.annotation.Required;
+import lombok.Data;
 
+@Data
 public class AnnotatedSpec {
   @JsonProperty("from-field")
   @JsonPropertyDescription("from-field-description")
@@ -37,6 +39,8 @@ public class AnnotatedSpec {
   private String singleDigit;
   private String nullable;
   private String defaultValue;
+  @Default("my-value2")
+  private String defaultValue2;
   @Required
   private boolean emptySetter;
   @Required